+++ episen-ms-basket/src/main/java/com/episen/basket/resource/ItemResource.java
@@ -2,6 +2,7 @@ package com.episen.basket.resource;
 
 
 import com.episen.basket.model.Item;
+import com.episen.basket.service.BasketService;
 import com.episen.basket.service.ItemService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
@@ -9,6 +10,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.text.ParseException;
 import java.util.List;
 
 @RestController
@@ -18,32 +20,56 @@ public class ItemResource {
     @Autowired
     private ItemService itemService;
 
+    @Autowired
+    private BasketService basketService;
+
     @GetMapping
-    public ResponseEntity<List<Item>> getAllItems() {
-        return new ResponseEntity<List<Item>>(itemService.getAllItems(), HttpStatus.OK);
+    public ResponseEntity<List<Item>> getAllItems(@RequestHeader(HttpHeaders.AUTHORIZATION) String token) throws ParseException {
+
+        String[] bearerToken = token.split("Bearer ");
+        String jwtToken = bearerToken[1];
+        if(!basketService.isTokenExpired(jwtToken) && itemService.isAdmin(jwtToken)){
+            return new ResponseEntity<List<Item>>(itemService.getAllItems(), HttpStatus.OK);
+        }
+        return new ResponseEntity<List<Item>>(HttpStatus.UNAUTHORIZED);
     }
 
     // ONLY ADMIN USER
     @PostMapping
-    public ResponseEntity<Object> add(@RequestBody Item item){
-        //TODO verifier if JwtToken is Admin before add Item
-        itemService.addItem(item);
-        return new ResponseEntity<Object>(item, HttpStatus.OK);
+    public ResponseEntity<Item> add(@RequestBody Item item, @RequestHeader(HttpHeaders.AUTHORIZATION) String token) throws ParseException {
+
+        String[] bearerToken = token.split("Bearer ");
+        String jwtToken = bearerToken[1];
+        if(!basketService.isTokenExpired(jwtToken) && itemService.isAdmin(jwtToken)){
+            itemService.addItem(item);
+            return new ResponseEntity<Item>(item, HttpStatus.CREATED);
+        }
+        return new ResponseEntity<Item>(HttpStatus.UNAUTHORIZED);
     }
 
     // ONLY ADMIN USER
     @PutMapping
-    public ResponseEntity<Object> update(@RequestBody Item itemToUpdate){
-        //TODO verifier if JwtToken is Admin before add Item
-        itemService.update(itemToUpdate);
-        return new ResponseEntity<Object>(itemToUpdate, HttpStatus.OK);
+    public ResponseEntity<Item> update(@RequestBody Item itemToUpdate, @RequestHeader(HttpHeaders.AUTHORIZATION) String token) throws ParseException {
+
+        String[] bearerToken = token.split("Bearer ");
+        String jwtToken = bearerToken[1];
+        if(!basketService.isTokenExpired(jwtToken) && itemService.isAdmin(jwtToken)){
+            itemService.update(itemToUpdate);
+            return new ResponseEntity<Item>(itemToUpdate, HttpStatus.OK);
+        }
+        return new ResponseEntity<Item>(HttpStatus.UNAUTHORIZED);
     }
 
     // ONLY ADMIN USER
     @DeleteMapping("{gtnToDelete}")
-    public ResponseEntity<Object> delete(@PathVariable Integer gtnToDelete){
-        //TODO verifier if JwtToken is Admin before add Item
-        itemService.delete(gtnToDelete);
-        return new ResponseEntity<Object>(gtnToDelete, HttpStatus.OK);
+    public ResponseEntity<Long> delete(@PathVariable Long gtnToDelete, @RequestHeader(HttpHeaders.AUTHORIZATION) String token) throws ParseException {
+
+        String[] bearerToken = token.split("Bearer ");
+        String jwtToken = bearerToken[1];
+        if(!basketService.isTokenExpired(jwtToken) && itemService.isAdmin(jwtToken)){
+            itemService.delete(gtnToDelete);
+            return new ResponseEntity<Long>(gtnToDelete, HttpStatus.OK);
+        }
+        return new ResponseEntity<Long>(HttpStatus.UNAUTHORIZED);
     }
 }
